项目重构

+++ 婚恋系统/Girl.h
@@ -26,6 +26,6 @@ private:
 	int age;
 	string name;
 	*/
-	int yanZhi;  //бежЕ
+	double yanZhi;  //бежЕ
 };
